refactor(dashboard): consolidate duplicate utility functions into src/lib/ modules
Extract 12 duplicated pure utility functions from 14 component files into
4 new canonical modules under dashboard/src/lib/:

- format-time.ts: relativeTime, formatElapsed, formatDuration, formatElapsedCompact, formatTimeAgo
- truncate.ts: trimText, truncate
- tone.ts: toneClass, chainStatusTone, sessionStatusTone, expiryTone, governanceToneClass
- status-label.ts: statusLabel, displayStatus, prettyJson

Original files re-export from lib/ to preserve existing import paths.
Net removal: ~345 lines of duplicated logic.
